web/requirements/requirements_linux.txt
Django==3.0.3
django-cors-headers==3.2.1
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
gunicorn==20.0.4
pycodestyle==2.5.0
pycparser==2.19
PyHamcrest==2.0.0
pylint==2.4.4
pyOpenSSL==19.1.0
pytz==2019.3
psycopg2-binary==2.8.4
service-identity==18.1.0
six==1.14.0
sqlparse==0.3.0
Twisted==19.10.0
txaio==20.1.1
typed-ast==1.4.1
uvicorn==0.11.2
uvloop==0.14.0
websockets==8.1
wrapt==1.11.2
zope.interface==4.7.1
djangorestframework-simplejwt==4.4.0
factory-boy==2.12.0 
Faker==4.0.1  
six==1.14.0 
django-phonenumber-field[phonenumbers]==4.0.0
django-storages==1.9.1
boto3==1.13.11
django-extra-fields==2.0.5
django-extensions==2.2.9
drf-yasg==1.17.1
django-guardian==2.2.0
django-filter==2.3.0
Pillow==2.2.1
python-dotenv==0.15.0





